requirements.txt
solara>=1.30.0
ipycanvas>=0.13.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
